requirements.txt
# LangGraph Orchestration
langgraph>=0.2.0
langchain-core>=0.3.0

# Core API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Audio/Video Processing
ffmpeg-python>=0.2.0
opencv-python-headless>=4.8.0
mediapipe>=0.10.8

# Speech & NLP
whisperx @ git+https://github.com/m-bain/whisperX.git
torch>=2.1.0
torchaudio>=2.1.0
silero-vad>=5.1

# Semantic Analysis
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0

# LLM (Azure AI Inference for GitHub Models)
azure-ai-inference>=1.0.0b1
azure-core>=1.30.0

# Utilities
aiofiles>=23.2.0
